Move tests in their own submodule
